requirements.txt
torch==1.0.1
torchvision==0.2.1
tqdm==4.26.0     
matplotlib==3.0.0
Pillow==5.2.0      
numpy==1.15.4
scikit-image==0.14.2           